Move ConsoleIntegrationTestTrait to console package
