+++ tests/TestCase/Command/PluginAssetsCommandsTest.php
@@ -19,10 +19,10 @@ namespace Cake\Test\TestCase\Command;
 use Cake\Command\Command;
 use Cake\Console\ConsoleIo;
 use Cake\Console\ConsoleOptionParser;
+use Cake\Console\TestSuite\ConsoleIntegrationTestTrait;
+use Cake\Console\TestSuite\StubConsoleOutput;
 use Cake\Core\Configure;
 use Cake\Filesystem\Filesystem;
-use Cake\TestSuite\ConsoleIntegrationTestTrait;
-use Cake\TestSuite\Stub\ConsoleOutput;
 use Cake\TestSuite\TestCase;
 use SplFileInfo;
 
@@ -111,7 +111,7 @@ class PluginAssetsCommandsTest extends TestCase
     {
         $this->loadPlugins(['TestTheme']);
 
-        $output = new ConsoleOutput();
+        $output = new StubConsoleOutput();
         $io = $this->getMockBuilder(ConsoleIo::class)
             ->setConstructorArgs([$output, $output, null, null])
             ->addMethods(['in'])
