+++ tests/TestCase/Console/Command/HelpCommandTest.php
@@ -17,9 +17,9 @@ declare(strict_types=1);
 namespace Cake\Test\TestCase\Console\Command;
 
 use Cake\Command\Command;
+use Cake\Console\TestSuite\ConsoleIntegrationTestTrait;
 use Cake\Core\Plugin;
 use Cake\Http\BaseApplication;
-use Cake\TestSuite\ConsoleIntegrationTestTrait;
 use Cake\TestSuite\TestCase;
 
 /**
