+++ tests/TestCase/Console/TestSuite/ConsoleIntegrationTestTraitTest.php
@@ -13,11 +13,11 @@ declare(strict_types=1);
  * @since         3.5.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Test\TestCase\TestSuite;
+namespace Cake\Test\TestCase\Console\TestSuite;
 
 use Cake\Command\Command;
-use Cake\TestSuite\ConsoleIntegrationTestTrait;
-use Cake\TestSuite\Stub\MissingConsoleInputException;
+use Cake\Console\TestSuite\ConsoleIntegrationTestTrait;
+use Cake\Console\TestSuite\MissingConsoleInputException;
 use Cake\TestSuite\TestCase;
 use PHPUnit\Framework\AssertionFailedError;
 use stdClass;
